refactored BaseChatMemory,ChatMessageHistory,get_prompt_input_key
