+++ docs/docs/modules/memory/chat_messages/index.mdx
@@ -13,7 +13,8 @@ This is a super lightweight wrapper that provides convenience methods for saving
 You may want to use this class directly if you are managing memory outside of a chain.
 
 ```python
-from langchain.memory import ChatMessageHistory
+
+from langchain_core.chat_history import ChatMessageHistory
 
 history = ChatMessageHistory()
