+++ docs/docs/modules/memory/types/summary.mdx
@@ -5,8 +5,9 @@ Conversation summary memory summarizes the conversation as it happens and stores
 Let's first explore the basic functionality of this type of memory.
 
 ```python
-from langchain.memory import ConversationSummaryMemory, ChatMessageHistory
+from langchain.memory import ConversationSummaryMemory
 from langchain.llms import OpenAI
+from langchain_core.chat_history import ChatMessageHistory
 ```
