+++ libs/community/langchain_community/chat_message_histories/__init__.py
@@ -43,8 +43,8 @@ from langchain_community.chat_message_histories.zep import ZepChatMessageHistory
 
 __all__ = [
     "AstraDBChatMessageHistory",
-    "ChatMessageHistory",
     "CassandraChatMessageHistory",
+    "ChatMessageHistory",
     "CosmosDBChatMessageHistory",
     "DynamoDBChatMessageHistory",
     "ElasticsearchChatMessageHistory",
