+++ libs/community/langchain_community/chat_message_histories/in_memory.py
@@ -1,21 +1,6 @@
-from typing import List
+from langchain_core.chat_history import ChatMessageHistory
 
-from langchain_core.chat_history import BaseChatMessageHistory
-from langchain_core.messages import BaseMessage
-from langchain_core.pydantic_v1 import BaseModel, Field
+__all__ = ["ChatMessageHistory"]
 
 
-class ChatMessageHistory(BaseChatMessageHistory, BaseModel):
-    """In memory implementation of chat message history.
 
-    Stores messages in an in memory list.
-    """
-
-    messages: List[BaseMessage] = Field(default_factory=list)
-
-    def add_message(self, message: BaseMessage) -> None:
-        """Add a self-created message to the store"""
-        self.messages.append(message)
-
-    def clear(self) -> None:
-        self.messages = []
