+++ libs/langchain/langchain/memory/chat_message_histories/in_memory.py
@@ -1,3 +1,3 @@
-from langchain_community.chat_message_histories.in_memory import ChatMessageHistory
+from langchain_core.chat_history import ChatMessageHistory
 
 __all__ = ["ChatMessageHistory"]
