+++ libs/langchain/langchain/memory/kg.py
@@ -4,6 +4,7 @@ from langchain_core.language_models import BaseLanguageModel
 from langchain_core.messages import BaseMessage, SystemMessage, get_buffer_string
 from langchain_core.prompts import BasePromptTemplate
 from langchain_core.pydantic_v1 import Field
+from langchain_core.utils.utils import get_prompt_input_key
 
 from langchain.chains.llm import LLMChain
 from langchain.graphs import NetworkxEntityGraph
@@ -13,7 +14,6 @@ from langchain.memory.prompt import (
     ENTITY_EXTRACTION_PROMPT,
     KNOWLEDGE_TRIPLE_EXTRACTION_PROMPT,
 )
-from langchain.memory.utils import get_prompt_input_key
 
 
 class ConversationKGMemory(BaseChatMemory):
