+++ libs/langchain/langchain/memory/vectorstore.py
@@ -3,11 +3,10 @@
 from typing import Any, Dict, List, Optional, Sequence, Union
 
 from langchain_core.documents import Document
+from langchain_core.memory import BaseMemory
 from langchain_core.pydantic_v1 import Field
+from langchain_core.utils.utils import get_prompt_input_key
 from langchain_core.vectorstores import VectorStoreRetriever
-
-from langchain.memory.chat_memory import BaseMemory
-from langchain.memory.utils import get_prompt_input_key
 
 
 class VectorStoreRetrieverMemory(BaseMemory):
